Feat/disable say delete (#19)
* - add say-auto-delete command (default to on)
-

* add 'nicknames' to help for `help ee`

+++ bot/src/commands/announce/say.ts
@@ -1,7 +1,7 @@
 import { Message } from 'discord.js'
 import { Command, CommandoClient, CommandoMessage } from 'discord.js-commando'
 import { say } from '../../lib/announce'
-import { DEFAULT_SPEED, DEFAULT_VOICE_ID, MAX_CHARS } from '../../lib/constants'
+import { DEFAULT_SAY_AUTODELETE, DEFAULT_SPEED, DEFAULT_VOICE_ID, MAX_CHARS, SAY_AUTO_DELETE } from '../../lib/constants'
 
 const MIN_CHARS = 2
 
@@ -49,15 +49,19 @@ export class SayCommand extends Command {
         DEFAULT_VOICE_ID,
       )
       const guildSpeed = message.guild.settings.get('speed', DEFAULT_SPEED)
+      const guildSayAutoDelete = message.guild.settings.get(SAY_AUTO_DELETE, DEFAULT_SAY_AUTODELETE)
 
       if (!voiceConnection) {
         return message.say(`You must be in the same voice channel as the bot.`)
       }
+
       // remove the message from the text-channel
-      try {
-        await message.delete()
-      } catch (err) {
-        console.error(`Error: Missing permission to delete this user's message`)
+      if(guildSayAutoDelete) {
+        try {
+          await message.delete()
+        } catch (err) {
+          console.error(`Error: Missing permission to delete this user's message`)
+        }
       }
 
       // announce
